Adiciona classe de teste.
