+++ app/src/main/java/br/ufg/inf/pes/healthhelp/dao/AbstractDAO.java
@@ -14,10 +14,10 @@ public abstract class AbstractDAO<T> implements InterfaceDAO<T>{
     private DatabaseCallback databaseCallback;
     private DatabaseReference databaseReference;
 
+
     public AbstractDAO(String TAG, String DATABASE_CHILD) {
         this.TAG = TAG;
         this.DATABASE_CHILD = DATABASE_CHILD;
-        this.databaseReference = FirebaseDatabase.getInstance().getReference();
     }
 
     public DatabaseCallback getDatabaseCallback() {
